Create tags and post tags
When you create posts, you'll be given the option to add tags to it. This will
allow posts to have many tags, which can be displayed on the post and help
with searching for posts. It also means tags can have many posts, so we can
display those posts on a tag page.

We'll also store the creator of the tag against the tag as 'creator' and let
users have many tags they've created.

Later we can let users have tags as skills, interests and/or their own style
of art.