Allow adding tags to posts
This adds a searchable select to post editor so you can add any existing tags
to a post for easier searching and tagging of data.

For now I've not yet been able to create tags when they don't exist using this
drop down, which is ideally how we'd like it setup so more tags are added to
the database at the point where a user needs them. I will work on this.