Allow users to like posts
This allows users to like and unlike a post (the heart is filled if they've
liked it).

The like link partial displays the like count and a list first names of the
first 3 users to like it that links to their profile, then a count of the
remaining users who have liked it.

Later we can do a popover with the full list of linked users who have liked a
post.